// src/applications/enums/application-status.enum.ts
export enum ApplicationStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  UNDER_REVIEW = 'UNDER_REVIEW',
  CORRECTION_REQUIRED = 'CORRECTION_REQUIRED',
  APPOINTMENT_SELECTION_REQUIRED = 'APPOINTMENT_SELECTION_REQUIRED',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  REINSPECTION_REQUIRED = 'REINSPECTION_REQUIRED',
  CANCELLED = 'CANCELLED',
  COMPLETED = 'COMPLETED',
}